feat: update readme and cleanup
