+++ README.md
@@ -1,25 +1,24 @@
 ![image](https://content.invisioncic.com/e290497/monthly_2024_03/new-cover.jpg.266c89d776592f674fd3a4133d3fb813.jpg)
-![Foundry v11](https://img.shields.io/badge/foundry-v11-green)
+![Foundry v12](https://img.shields.io/badge/foundry-v12-green)
 
 # Metanthropes - Early Access
 
 This is the official Metanthropes System for Foundry VTT.
 
--   This project is in no way affiliated with or supported by Foundry VTT.
--   For more information about Metanthropes visit [our website](https://metanthropes.com)
--   We would love to hear your opinion and feedback for our Early Access Demo build. Please [join us on Discord](https://metanthropes.com/discord)
+-   For more information about Metanthropes visit [our website.](https://metanthropes.com)
+-   We would love to hear your opinion and feedback! Please [join us on Discord!](https://metanthropes.com/discord)
 
 ## Installation
 
 This System is a work in progress and is not ready for general use. This Early Access build is for demonstration purposes only.
 
-To install this System, please search for 'Metanthropes' in Foundry's System Search box, alternatively you may use the following link when selecting to install manually via a .JSON manifest file:
+To use this System, please search for 'Metanthropes' in Foundry's System Search box and click Install. Alternatively you may use the following link when selecting to install manually via a .JSON manifest file:
 
 	 https://github.com/legitamine/metanthropes/releases/latest/download/system.json
 
 ## Supported Modules
 
-This System aims to utilize the latest and best in Foundry VTT. Currently offers support for Foundry VTT version 11. Latest supported version is v11.315.
+This System aims to utilize the latest and best in Foundry VTT. Currently offers support for Foundry VTT version 12. Latest supported version is v12.331.
 
 To get the best experience while playing Metanthropes, we recommend to install the following Modules:
 
@@ -27,32 +26,38 @@ To get the best experience while playing Metanthropes, we recommend to install t
 -   [Carousel Combat Tracker](https://foundryvtt.com/packages/combat-tracker-dock)
 -   [Drag Ruler](https://foundryvtt.com/packages/drag-ruler)
 
-A more comprehensive list of supported & recommended Modules will be made available soon™
+A more comprehensive list of supported & recommended Modules will be made available soon™.
 
 ## Planned Features
 
 Briefly, we are working torwards the following milestones:
 
--   Officially Support the full Ruleset for Metanthropes, including automating the vast majority of the game mechanics
--   Make Combat as immersive as possible, with auto-applying damage/healing, targeting via measured templates and controlling all types of effects to Actors
+-   Officially Support the full Ruleset for Metanthropes, including automating the vast majority of the game mechanics.
+-   Make Combat as immersive as possible, with auto-applying damage/healing, targeting via measured templates and controlling all types of effects to Actors.
 -   Further automate the activation of Metapowers, including unique audio and visual effects
--   Take advantage of upcoming v12 Features, such as the Scene Zones to help futher automate the application of Metapower Effects
--   Synchronize your Protagonist details with our website for off-line viewing and editing between Sessions, without having to open Foundry VTT
+-   Synchronize your Protagonist details with our website for off-line viewing and editing between Sessions, without having to open Foundry VTT.
 
-For further info, check out our [planned features and milestones](https://github.com/Legitamine/metanthropes/projects?query=is%3Aopen) for this project and also check out the overall [roadmap for Metanthropes](https://www.metanthropes.com/stratagem/projects/1-metanthropes-roadmap/)
+For further info, check out our [planned features and milestones](https://github.com/Legitamine/metanthropes/projects?query=is%3Aopen) for this project and also check out the overall [roadmap for Metanthropes](https://www.metanthropes.com/stratagem/projects/1-metanthropes-roadmap/).
 
 ## Latest Changes
 
-For the latest changes for each build you may check out the [detailed Change Log](https://github.com/Legitamine/metanthropes/blob/main/CHANGELOG.md)
+For the latest changes for each build you may check out the [detailed Change Log](https://github.com/Legitamine/metanthropes/blob/main/CHANGELOG.md).
 
-You can also view the archives of past releases, kept for posterity, [here](https://github.com/Legitamine/metanthropes/blob/main/CHANGELOGARCHIVES.md)
+You can also view the archives of past releases, kept for posterity, [here](https://github.com/Legitamine/metanthropes/blob/main/CHANGELOGARCHIVES.md).
 
-Our eternal gratitude goes to the amazing community of FVTT Developers <3
+## Shoutouts & Thanks
+
+Our eternal gratitude goes to the amazing community of FVTT Developers ❤️!
 
 Shoutout especially to the #system-development channel on the FVTT Discord server & extra-special thanks to the following people:
- - @TyphonJS (Michael) for the code used in the metaLogDocument function
- - @Zhell (zhell9201) for the overall help, guidance & suggestions
- - @Mana (manaflower) for the overall help, guidance & CSS wizardry
- - @ChaosOS for the overall help & CLI guidance
- - @mxzf for the overall help & patience & guidance & the amazing website of FVTT resources
- - @asacolips for the Boilerplate code, without which I would have been completely lost and probably would have given up very early on
+ - @TyphonJS (Michael) for the code used in the metaLogDocument function.
+ - @Zhell (zhell9201) for the overall help, guidance & suggestions.
+ - @Mana (manaflower) for the overall help, guidance & CSS wizardry.
+ - @mxzf for the overall help & patience & guidance & the amazing website of FVTT resources!
+ - @ChaosOS for the overall help & CLI guidance and for the many contributions to the Wiki!
+ - @asacolips for the Boilerplate code, without which I would have been completely lost and probably would have given up very early on!!
+
+Special shoutouts and huge thanks to the lovely community of Metanthropes, for their continued support and feedback:
+ - @RXOliver and the gang over at the THC Discord for the playtesting and suggestions!
+ - @MB & party for amazing feedback and suggestions, and for being the first to contribute specific code suggestions for the project!
+ - Finally a special shoutout to my buddy @aMUSiC, who has guided me to avoid many pitfalls and for code-review assistance & overall mentoring!!
